Defaul list controller ok, test comentados de default list

+++ app/Http/Controllers/api/v1/DefaultListSecretaryController.php
@@ -16,14 +16,22 @@ class DefaultListSecretaryController extends Controller
     public function __construct()
     {
         $this->middleware('auth');
-        $this->middleware('checkrolesapi:SECRETARY');
+        $this->middleware('checkrolesapi:PRESIDENT|COORDINATOR|REGISTER_COORDINATOR|SECRETARY|STUDENT');
     }
 
     public function list()
     {
         //$instance = \Instantiation::instance();
-        $defaultlists = auth('api')->user()->secretary->default_lists()->get();
-        return $defaultlists;
+
+        if(auth('api')->user()->secretary) {
+            $defaultlists = auth('api')->user()->secretary->default_lists()->get();
+            return $defaultlists;
+        } else {
+            return response()->json([
+                'success' => false,
+                'message' => 'El usuario no tiene permisos.'
+            ], 403);
+        }
     }
 
     /*public function create()
@@ -38,34 +46,40 @@ class DefaultListSecretaryController extends Controller
 
     public function new(Request $request)
     {
-
         // $instance = \Instantiation::instance();
 
-        $validatedData = $request->validate([
-            'name' => 'required|max:255',
-            'users' => 'required|array|min:1',
-        ]);
+        if(auth('api')->user()->secretary) {
+            $validatedData = $request->validate([
+                'name' => 'required|max:255',
+                'users' => 'required|array|min:1',
+            ]);
 
-        $secretary = Auth::user()->secretary;
+            $secretary = Auth::user()->secretary;
 
-        $defaultlist = DefaultList::create([
-            'name' => $request->input('name'),
-            'secretary_id' => $secretary->id
-        ]);
+            $defaultlist = DefaultList::create([
+                'name' => $request->input('name'),
+                'secretary_id' => $secretary->id
+            ]);
 
-        $defaultlist->save();
+            $defaultlist->save();
 
-        // Asociamos los usuarios a la reunión
-        $users_ids = $request->input('users',[]);
+            // Asociamos los usuarios a la reunión
+            $users_ids = $request->input('users',[]);
 
-        foreach($users_ids as $user_id)
-        {
+            foreach($users_ids as $user_id)
+            {
 
-            $user = User::find($user_id);
-            $defaultlist->users()->attach($user);
+                $user = User::find($user_id);
+                $defaultlist->users()->attach($user);
 
+            }
+            return $defaultlist;
+        } else {
+            return response()->json([
+                'success' => false,
+                'message' => 'El usuario no tiene permisos.'
+            ], 403);
         }
-        return $defaultlist;
     }
 
     /*public function edit($instance,$id)
@@ -92,25 +106,34 @@ class DefaultListSecretaryController extends Controller
 
         $defaultlist = DefaultList::find($id);
         $defaultlist->name = $request->input('name');
-        $defaultlist->save();
 
-        // Asociamos los usuarios a la reunión
-        $users_ids = $request->input('users',[]);
+        $userid = $defaultlist->secretary->user->id;
 
-        // eliminamos usuarios antiguos de la reunión
-        foreach($defaultlist->users as $user)
-        {
-            $defaultlist->users()->detach($user);
+        if(auth('api')->user()->secretary && auth('api')->id() == $userid) {
+            $defaultlist->save();
+
+            // Asociamos los usuarios a la reunión
+            $users_ids = $request->input('users',[]);
+
+            // eliminamos usuarios antiguos de la reunión
+            foreach($defaultlist->users as $user)
+            {
+                $defaultlist->users()->detach($user);
+            }
+
+            // agregamos los usuarios nuevos de la reunión
+            foreach($users_ids as $user_id)
+            {
+                $user = User::find($user_id);
+                $defaultlist->users()->attach($user);
+            }
+            return $defaultlist;
+        }  else {
+            return response()->json([
+                'success' => false,
+                'message' => 'El usuario no tiene permisos.'
+            ], 403);
         }
-
-        // agregamos los usuarios nuevos de la reunión
-        foreach($users_ids as $user_id)
-        {
-            $user = User::find($user_id);
-            $defaultlist->users()->attach($user);
-        }
-
-        return $defaultlist;
     }
 
     public function remove(Request $request, $instance, $id)
@@ -118,9 +141,16 @@ class DefaultListSecretaryController extends Controller
         //$instance = \Instantiation::instance();
 
         $defaultlist = DefaultList::find($id);
+        $userid = $defaultlist->secretary->user->id;
 
-        $defaultlist->delete();
-
-        return response()->json('Lista por defecto eliminada con éxito');
+        if(auth('api')->user()->secretary && auth('api')->id() == $userid) {
+            $defaultlist->delete();
+            return response()->json('Lista por defecto eliminada con éxito');
+        }  else {
+            return response()->json([
+                'success' => false,
+                'message' => 'El usuario no tiene permisos.'
+            ], 403);
+        }
     }
 }
